+++ src/main/java/id/swhp/javaee/soteria/application/exception/InvalidCredentialException.java
@@ -1,0 +1,10 @@
+package id.swhp.javaee.soteria.application.exception;
+
+/**
+ *
+ * @author Sukma Wardana
+ * @since 1.0
+ */
+public class InvalidCredentialException extends EnterpriseException {
+
+}
