+++ src/main/java/id/swhp/javaee/soteria/business/exception/InvalidPasswordException.java
@@ -1,0 +1,12 @@
+package id.swhp.javaee.soteria.business.exception;
+
+import id.swhp.javaee.soteria.application.exception.EnterpriseException;
+
+/**
+ *
+ * @author Sukma Wardana
+ * @since 1.0
+ */
+public class InvalidPasswordException extends EnterpriseException {
+
+}
