--- src/main/java/id/swhp/javaee/soteria/business/exception/boundary/InvalidPasswordException.java
@@ -1,10 +1,0 @@
-package id.swhp.javaee.soteria.business.exception.boundary;
-
-/**
- *
- * @author Sukma Wardana
- * @since 1.0
- */
-public class InvalidPasswordException extends BusinessException {
-
-}
