+++ src/main/resources/db/schema_postgres10.sql
@@ -1,7 +1,6 @@
--- DDL for PostgreSQL up to 9 and MariaDB/MySQL 5 or above
--- For PostgreSQL 10 and above please use schema_postgres10.sql
+-- DDL for PostgreSQL 10 or above
 CREATE TABLE account(
-   id SERIAL PRIMARY KEY,
+   id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
    username VARCHAR(100) UNIQUE NOT NULL,
    password VARCHAR(255) NOT NULL,
    email VARCHAR(100) UNIQUE NOT NULL,
@@ -9,7 +8,7 @@ CREATE TABLE account(
 );
 
 CREATE TABLE token(
-   id SERIAL PRIMARY KEY,
+   id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
    account_id INTEGER REFERENCES account(id),
    token_hash VARCHAR(255) NOT NULL,
    token_TYPE VARCHAR(100),
